requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
groq
Werkzeug==2.3.7
python-dotenv==1.0.0